+++ web/src/stream_sort.js
@@ -1,5 +1,8 @@
+import * as settings_config from "./settings_config";
 import * as stream_data from "./stream_data";
+import * as stream_topic_history from "./stream_topic_history";
 import * as sub_store from "./sub_store";
+import {user_settings} from "./user_settings";
 import * as util from "./util";
 
 let previous_pinned;
@@ -8,6 +11,13 @@ let previous_dormant;
 let previous_muted_active;
 let previous_muted_pinned;
 let all_streams = [];
+
+// Because we need to check whether we are filtering inactive streams
+// in a loop over all streams to render the left sidebar, and the
+// definition of demote_inactive_streams involves how many streams
+// there are, we maintain this variable as a cache of the calculation
+// to avoid making left sidebar rendering a quadratic operation.
+let filter_out_inactives = false;
 
 export function get_streams() {
     const sorted_streams = all_streams.map((stream_id) =>
@@ -26,6 +36,45 @@ function compare_function(a, b) {
     return util.strcmp(stream_name_a, stream_name_b);
 }
 
+export function set_filter_out_inactives() {
+    if (
+        user_settings.demote_inactive_streams ===
+        settings_config.demote_inactive_streams_values.automatic.code
+    ) {
+        filter_out_inactives = stream_data.num_subscribed_subs() >= 30;
+    } else if (
+        user_settings.demote_inactive_streams ===
+        settings_config.demote_inactive_streams_values.always.code
+    ) {
+        filter_out_inactives = true;
+    } else {
+        filter_out_inactives = false;
+    }
+}
+
+// Exported for access by unit tests.
+export function is_filtering_inactives() {
+    return filter_out_inactives;
+}
+
+export function has_recent_activity(sub) {
+    if (!filter_out_inactives || sub.pin_to_top) {
+        // If users don't want to filter inactive streams
+        // to the bottom, we respect that setting and don't
+        // treat any streams as dormant.
+        //
+        // Currently this setting is automatically determined
+        // by the number of streams.  See the callers
+        // to set_filter_out_inactives.
+        return true;
+    }
+    return stream_topic_history.stream_has_topics(sub.stream_id) || sub.newly_subscribed;
+}
+
+export function has_recent_activity_but_muted(sub) {
+    return has_recent_activity(sub) && sub.is_muted;
+}
+
 export function sort_groups(streams, search_term) {
     const stream_id_to_name = (stream) => sub_store.get(stream).name;
     // Use -, _ and / as word separators apart from the default space character
@@ -38,7 +87,7 @@ export function sort_groups(streams, search_term) {
     );
 
     function is_normal(sub) {
-        return stream_data.is_active(sub);
+        return has_recent_activity(sub);
     }
 
     const pinned_streams = [];
@@ -137,4 +186,8 @@ export function next_stream_id(stream_id) {
     }
 
     return maybe_get_stream_id(i + 1);
+}
+
+export function initialize() {
+    set_filter_out_inactives();
 }
