+++ web/src/typeahead_helper.js
@@ -11,6 +11,7 @@ import * as people from "./people";
 import * as pm_conversations from "./pm_conversations";
 import * as recent_senders from "./recent_senders";
 import * as stream_data from "./stream_data";
+import * as stream_sort from "./stream_sort";
 import * as user_groups from "./user_groups";
 import * as user_status from "./user_status";
 import * as util from "./util";
@@ -430,8 +431,8 @@ function activity_score(sub) {
         if (sub.pin_to_top) {
             stream_score += 2;
         }
-        // Note: A pinned stream may accumulate a 3rd point if it is active
-        if (stream_data.is_active(sub)) {
+        // Note: A pinned stream may accumulate a 3rd point if it has recent activity.
+        if (stream_sort.has_recent_activity(sub)) {
             stream_score += 1;
         }
     }
